-Fixed compiler warnings, memset warnings, and moved preset C file creation inside DEVELOPMENT conditionals -Updated windows UI for modlines, mainWindow and Settings -Implemented root level content folder that gets copied into package dirs -Updated repo readme build instructions -Fixed issues with bootloader install, app relaunch -Updated build process and Windows UI in EZ editor -Added windows installer/deploy method from other Qt editors, fixed windows icons post install

+++ standalone/advanced/hosted/navdatacooker.cpp
@@ -22,11 +22,11 @@ NavDataCooker::NavDataCooker(QWidget *parent) :
     offThreshW = 5;
 
     //Set temporarily here until settings are hooked up
-    globalGain = 1.00;
-    navNGain = 1.1;
-    navSGain = 1.1;
-    navEGain = 1.0;
-    navWGain = 1.0;
+    globalGain = (float)1.00;
+    navNGain = (float)1.1;
+    navSGain = (float)1.1;
+    navEGain = (float)1.0;
+    navWGain = (float)1.0;
 
     yAccel = 10;
